corrected file option bug

+++ src/extended/align_free_score.c
@@ -251,7 +251,7 @@ double **calc_fscore(const GtEncseq *encseq_first,
 GtWord **calc_edist(const GtEncseq *encseq_first,
                     const GtEncseq *encseq_second,
                     GtStr *scorematrix,
-                    int indelscore,
+                    double indelscore,
                     GtError *err)
 {
   GtScoreMatrix *smatrix;
@@ -345,8 +345,8 @@ GtWord **calc_edist(const GtEncseq *encseq_first,
 GtWord **calc_edist_affine(const GtEncseq *encseq_first,
                            const GtEncseq *encseq_second,
                            GtStr *scorematrix,
-                           int gapopen,
-                           int gapextend,
+                           double gapopen,
+                           double gapextend,
                            GtError *err)
 {
   GtScoreMatrix *smatrix;
